rewrite all components/Profile files
