+++ mwdb/web/src/commons/auth/context.tsx
@@ -1,5 +1,5 @@
-import { AuthContextValues } from "@mwdb-web/types/types";
 import React from "react";
+import { AuthContextValues } from "@mwdb-web/types/context";
 
 export const AuthContext = React.createContext<AuthContextValues>(
     {} as AuthContextValues
