+++ mwdb/web/src/commons/config/context.tsx
@@ -1,4 +1,4 @@
-import { ConfigContextValues } from "@mwdb-web/types/types";
+import { ConfigContextValues } from "@mwdb-web/types/context";
 import React from "react";
 
 export const ConfigContext = React.createContext<ConfigContextValues>(
