+++ mwdb/web/src/commons/ui/DateString.tsx
@@ -1,8 +1,11 @@
 type Props = {
-    date: string;
+    date?: string | Date;
 };
 
 export default function DateString(props: Props) {
+    if (!props.date) {
+        return <></>;
+    }
     const date = props.date;
     const d = new Date(date);
     return <span>{date != null ? d.toUTCString() : "(never)"}</span>;
