+++ mwdb/web/src/commons/ui/ShowIf.tsx
@@ -1,8 +1,8 @@
 type Props = {
     condition: boolean;
-    children: React.ReactNode;
+    children: JSX.Element;
 };
 
 export function ShowIf({ condition, children }: Props) {
-    return condition ? children : [];
+    return condition ? children : <></>;
 }
